UPDATE: accept float and update column types

+++ app/controllers/gradeController.php
@@ -92,7 +92,7 @@ class GradeController extends Controller{
         $gradeSchemeId = $curriculumModel->schoolLevelGradeSchemeInfo($level_id);
         $gradeScheme = $gradeModel->gradeSchemeDetails((int)$gradeSchemeId['grade_scheme_id'] ?? 0);
 
-        $this->view->pass_threshold = $gradeScheme ? ((int)$gradeScheme[0]->pass_threshold ?? 0) : false;
+        $this->view->pass_threshold = $gradeScheme ? ((float)$gradeScheme[0]->pass_threshold ?? 0) : false;
         $this->view->targetSection = $section_id;
         $this->view->studentGrades = $studentGrades; //Contains all the grade this section have. Each array represents a student
         $this->view->studentSection = $studentSection;
